Add files via upload
Updating launch file to simulate the map and removed the lasers (temporarily, in line 26 )
Updated the fwd_kinematics node with subscriber data being recorded in a global variable. Tried adding the kinematics. v_x and w_z but its not working as of now.
Simulation still runs with jsp = True